modules/removal_export_pipeline/pyproject.toml
[tool.poetry]
name = "removal-export-pipeline"
version = "0.1.0"
description = "poetry for removal_export pipeline LengoAI"
authors = ["MihaiDavid05 <davidmihai98@yahoo.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
absl-py = "2.1.0"
aiofiles = "23.2.1"
jinja2 = "^3.1.2"
aiohappyeyeballs = "2.4.3"
aiohttp = "3.10.10"
annotated-types = "0.7.0"
antlr4-python3-runtime = "4.9.3"
anyio = "4.4.0"
asttokens = "2.4.1"
astunparse = "1.6.3"
attrs = "24.2.0"
build = "1.2.1"
cachecontrol = "0.14.0"
cachetools = "5.5.0"
certifi = "2024.7.4"
charset-normalizer = "3.3.2"
cleo = "2.1.0"
click = "8.1.7"
colorama = "0.4.6"
comm = "0.2.2"
contourpy = "1.2.1"
crashtest = "0.4.1"
cycler = "0.12.1"
debugpy = "1.8.7"
decorator = "5.1.1"
defusedxml = "0.7.1"
deprecated = "1.2.14"
distlib = "0.3.8"
dulwich = "0.21.7"
einops = "0.8.0"
fastapi = "0.112.1"
fastjsonschema = "2.20.0"
ffmpy = "0.4.0"
gcsfs = "2024.10.0"
geopandas = "1.0.1"
gradio = "4.42.0"
hydra-core = "1.3.2"
ipython = "8.29.0"
google-cloud = "0.34.0"
google-auth = "2.34.0"
google-cloud-logging = "3.11.2"
google-cloud-storage = "2.18.2"
google-cloud-core = "2.4.1"
keras = "3.6.0"
matplotlib = "3.9.2"
opencv-python = "^4.10.0.84"
pydantic = "^2.9.2"
requests = "^2.32.3"
pyyaml = "^6.0.2"
scikit-learn = "1.5.1"
scipy = "^1.14.1"
seaborn = "0.13.2"
shapely = "2.0.6"
supervision = "0.22.0"
tqdm = "4.66.5"
networkx = "^3.4.2"
pyshp = "^2.3.1"
python-dotenv = "^1.0.1"
poselib = "^2.0.4"
pycolmap = "0.6.1"
kornia = "0.7.3"
kornia_rs = "0.1.7"
unidecode = "^1.3.8"
XlsxWriter ="^3.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
